+++ packages/plugin-accordion/web/src/toolbar/insert-buttons.ts
@@ -1,14 +1,15 @@
 import { DEFAULTS } from '../accordion-component';
 import { TOOLBARS, BUTTON_TYPES } from 'wix-rich-content-editor-common';
 import InsertPluginIcon from '../icons/InsertPluginIcon';
-import { CreateInsertButtons, TranslationFunction, PluginConfig } from 'wix-rich-content-common';
+import { CreateInsertButtons, TranslationFunction } from 'wix-rich-content-common';
+import { AccordionPluginEditorConfig } from '../types';
 
 const createInsertButtons: CreateInsertButtons = ({
   t,
   settings,
 }: {
   t: TranslationFunction;
-  settings: PluginConfig;
+  settings: AccordionPluginEditorConfig;
   isMobile: boolean;
 }) => {
   const icon = settings?.toolbar?.icons?.InsertPluginButtonIcon || InsertPluginIcon;