+++ packages/plugin-commons/web/src/Base/createBaseComponent.jsx
@@ -9,11 +9,6 @@ import {
   textWrapClassName,
   createHocName,
 } from 'wix-rich-content-common';
-import {
-  pluginsWithoutBorderOnHover,
-  pluginsWithoutBorderOnFocus,
-  pluginsWithoutPointerEventsOnFocus,
-} from '../consts';
 import styles from 'wix-rich-content-editor-common/dist/statics/styles/general.scss';
 import rtlIgnoredStyles from 'wix-rich-content-common/dist/statics/styles/general.rtlignore.scss';
 
@@ -26,7 +21,6 @@ const DEFAULTS = Object.freeze({
 
 const createBaseComponent = ({
   PluginComponent,
-  type,
   theme,
   settings,
   pubsub,
@@ -48,6 +42,8 @@ const createBaseComponent = ({
   anchorTarget,
   relValue,
   renderInnerRCE,
+  noPluginBorder,
+  noPointerEventsOnFocus,
 }) => {
   return class WrappedComponent extends Component {
     static propTypes = {
@@ -274,8 +270,11 @@ const createBaseComponent = ({
       const { width: currentWidth, height: currentHeight } = componentData.config || {};
       const { width: initialWidth, height: initialHeight } = settings || {};
       const isEditorFocused = selection.getHasFocus();
+      const isOneBlockSelected = selection.getStartKey() === selection.getEndKey();
+      const isPartOfSelection = blockProps.isFocused && !isOneBlockSelected;
+      blockProps.isFocused = blockProps.isFocused && isOneBlockSelected && isEditorFocused;
+
       const { isFocused } = blockProps;
-      const isActive = isFocused && isEditorFocused;
 
       const classNameStrategies = compact([
         PluginComponent.alignmentClassName || alignmentClassName,
@@ -284,11 +283,13 @@ const createBaseComponent = ({
         PluginComponent.customClassName,
       ]).map(strategy => strategy(this.state.componentData, theme, this.styles, isMobile));
 
+      const hasFocus = isFocused ? !noPluginBorder : isPartOfSelection;
+
       const ContainerClassNames = classNames(
         this.styles.pluginContainer,
         theme.pluginContainer,
         theme.pluginContainerWrapper,
-        pluginsWithoutBorderOnHover.includes(type) && this.styles.noBorderOnHover,
+        !isPartOfSelection && noPluginBorder && this.styles.noBorder,
         {
           [this.styles.pluginContainerMobile]: isMobile,
           [theme.pluginContainerMobile]: isMobile,
@@ -297,17 +298,16 @@ const createBaseComponent = ({
         classNameStrategies,
         className || '',
         {
-          [this.styles.hasFocus]: isActive && !pluginsWithoutBorderOnFocus.includes(type),
-          [theme.hasFocus]: isActive,
+          [this.styles.hasFocus]: hasFocus,
+          [theme.hasFocus]: hasFocus,
+          [this.styles.hideTextSelection]: !isFocused,
         }
       );
 
       const overlayClassNames = classNames(
         this.styles.overlay,
         theme.overlay,
-        isFocused &&
-          pluginsWithoutPointerEventsOnFocus.includes(type) &&
-          this.styles.noPointerEvents
+        isFocused && noPointerEventsOnFocus && this.styles.noPointerEventsOnFocus
       );
 
       const sizeStyles = {
@@ -348,7 +348,7 @@ const createBaseComponent = ({
           role="none"
           style={sizeStyles}
           className={ContainerClassNames}
-          data-focus={isActive}
+          data-focus={isFocused}
           onDragStart={this.onDragStart}
           onContextMenu={this.handleContextMenu}
           {...decorationProps}