+++ packages/plugin-file-upload/web/src/toolbar/insert-buttons.tsx
@@ -1,14 +1,15 @@
 import { TOOLBARS, INSERT_PLUGIN_BUTTONS, BUTTON_TYPES } from 'wix-rich-content-editor-common';
 import { DEFAULTS } from '../file-upload-component';
 import { InsertPluginIcon } from '../icons';
-import { CreateInsertButtons, TranslationFunction, PluginConfig } from 'wix-rich-content-common';
+import { CreateInsertButtons, TranslationFunction } from 'wix-rich-content-common';
+import { FilePluginEditorConfig } from '../types';
 
 const createInsertButtons: CreateInsertButtons = ({
   settings,
   t,
 }: {
   t: TranslationFunction;
-  settings: PluginConfig;
+  settings: FilePluginEditorConfig;
 }) => {
   const icon = settings?.toolbar?.icons?.InsertPluginButtonIcon || InsertPluginIcon;
   return [