+++ packages/plugin-giphy/web/src/constants.ts
@@ -3,7 +3,6 @@ import { ModalStyles } from 'wix-rich-content-common';
 export const PAGE_SIZE = 25;
 export const WAIT_INTERVAL = 300;
 export const SEARCH_TYPE = 'gifs';
-export const GIPHY_TYPE = 'wix-draft-plugin-giphy';
 
 export const DEFAULTS = Object.freeze({
   config: {