+++ packages/template-atomic-plugin/web/src/createYourPluginNamePlugin.ts
@@ -1,10 +1,10 @@
 import createToolbar from './toolbar/createToolbar';
 import { Component, DEFAULTS } from './yourDpluginDname-component';
-import { YOUR_PLUGIN_NAME_TYPE } from './types';
+import { YOUR_PLUGIN_NAME_TYPE, YourPluginNamePluginEditorConfig } from './types';
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
 import { CreatePluginFunction } from 'wix-rich-content-common';
 
-const createYourPluginNamePlugin: CreatePluginFunction = config => {
+const createYourPluginNamePlugin: CreatePluginFunction<YourPluginNamePluginEditorConfig> = config => {
   const { helpers, t, [YOUR_PLUGIN_NAME_TYPE]: settings = {}, isMobile, ...rest } = config;
 
   return createBasePlugin({