+++ CHANGELOG.md
@@ -16,10 +16,70 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `common`
+  - [#1761](https://github.com/wix-incubator/rich-content/pull/1761) customStyles: keep `.text` color as `unset` if custom paragraph color was not defined.
 
 </details>
 <hr/>
 
+## 8.4.8 (Nov 12, 2020)
+### :bug: Bug Fix
+- `headings`
+  - [#1752](https://github.com/wix-incubator/rich-content/pull/1752) fix headings drop down customization & changed `dropDownOptions` prop to `customHeadings`
+
+### :house: Internal
+- `general`
+  - [#1747](https://github.com/wix-incubator/rich-content/pull/1747) defined API for all editor & viewer plugins
+
+## 8.4.7 (Nov 12, 2020)
+### :house: Internal
+- `viewer`
+  - [#1750](https://github.com/wix-incubator/rich-content/pull/1750) bundle size optimzation: editor translations are excluded from bundle
+  
+### :bug: Bug Fix
+- `general`
+  - [#1749](https://github.com/wix-incubator/rich-content/pull/1749) ui fix for error display in galleries and images
+- `editor`
+  - [#1754](https://github.com/wix-incubator/rich-content/pull/1754) fix toast styles on mobile
+
+
+## 8.4.6 (Nov 11, 2020)
+### :bug: Bug Fix
+- `file-upload`
+  - [#1717](https://github.com/wix-incubator/rich-content/pull/1717) cursor pointer on hover
+- `editor`
+  - [#1746](https://github.com/wix-incubator/rich-content/pull/1746) media BI object properties fix
+  
+### :house: Internal
+- `editor`
+  - [#1756](https://github.com/wix-incubator/rich-content/pull/1756) fix `onMediaUploadEnd` BI data arguments
+
+## 8.4.5 (Nov 11, 2020)
+### :bug: Bug Fix
+- `common`
+  - [#1743](https://github.com/wix-incubator/rich-content/pull/1743) customStyles `list` inherits `p` values
+
+## 8.4.4 (Nov 10, 2020)
+### :bug: Bug Fix
+- `plugin-commons`
+  - [#1721](https://github.com/wix-incubator/rich-content/pull/1721) fix `isFocused = true` prop value when plugin is selected (not focused)
+- `accordion`
+  - [#1721](https://github.com/wix-incubator/rich-content/pull/1721) disable text selection when accordion is selected (not focused)
+- `common`
+  - [#1742](https://github.com/wix-incubator/rich-content/pull/1742) unset default `.text` fontSize
+
+### :house: Internal
+- `vertical-embed`
+  - [#1736](https://github.com/wix-incubator/rich-content/pull/1736) update mocks and add translations
+- `gallery`
+  - [#1739](https://github.com/wix-incubator/rich-content/pull/1739) bump pro gallery version to 2.2.16
+
+## 8.4.3 (Nov 5, 2020)
+### :bug: Bug Fix
+- `viewer`
+  - [#1731](https://github.com/wix-incubator/rich-content/pull/1731) fix justify in safari/firefox (adjust previous fix for the new dom structure)
+  
 ## 8.4.2 (Nov 5, 2020)
 ### :rocket: New Feature
 - `ricos-common`